+++ package.json
@@ -70,6 +70,7 @@
         "file-loader": "1.1.5",
         "fs-extra": "3.0.1",
         "html-webpack-plugin": "2.29.0",
+        "immer": "^2.0.0",
         "isomorphic-style-loader": "^4.0.0",
         "jest": "20.0.4",
         "lodash": "^4.17.11",
